fix: corregir tipo de int a long en UsuarioRepository

+++ src/main/java/com/wilsontut/kinalapp/repository/UsuarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
     public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
-    List<Usuario> findByEstado(int estado);
+    List<Usuario> findByEstado(long estado);
 }
